create a types file and move data types into this
