+++ xinfer/transformers/vlrm_blip2.py
@@ -1,6 +1,7 @@
 import torch
 
-from ..model_registry import ModelInputOutput, register_model
+from ..model_registry import register_model
+from ..types import ModelInputOutput
 from .blip2 import BLIP2
