+++ xinfer/types.py
@@ -1,0 +1,19 @@
+from dataclasses import dataclass
+from enum import Enum
+
+
+class ModelInputOutput(Enum):
+    IMAGE_TO_TEXT = "image --> text"
+    IMAGE_TEXT_TO_TEXT = "image-text --> text"
+    TEXT_TO_TEXT = "text --> text"
+    IMAGE_TO_BOXES = "image --> boxes"
+    IMAGE_TO_CATEGORIES = "image --> categories"
+    IMAGE_TO_MASKS = "image --> masks"
+    IMAGE_TO_POINTS = "image --> points"
+
+
+@dataclass
+class ModelInfo:
+    id: str
+    implementation: str
+    input_output: ModelInputOutput
