+++ xinfer/ultralytics/yolov10.py
@@ -1,4 +1,5 @@
-from ..model_registry import ModelInputOutput, register_model
+from ..model_registry import register_model
+from ..types import ModelInputOutput
 from .ultralytics_model import UltralyticsModel
