+++ xinfer/viz.py
@@ -3,8 +3,9 @@ import json
 import gradio as gr
 
 from .core import create_model
-from .model_registry import ModelInputOutput, model_registry
+from .model_registry import model_registry
 from .models import BaseModel
+from .types import ModelInputOutput
 
 
 def launch_gradio(model: BaseModel, **gradio_launch_kwargs):
